requirements.txt
et-xmlfile==1.1.0
lxml==4.7.1
openpyxl==3.0.9
pdf2image==1.16.0
Pillow==9.0.1
pytesseract==0.3.8
python-docx==0.8.11
python-pptx==0.6.21
XlsxWriter==3.0.2
